+++ pytest/tools/prober/prober_util.py
@@ -1,0 +1,80 @@
+#!/usr/bin/env python3
+
+import sys
+import json
+import time
+import pathlib
+import requests
+
+sys.path.append(str(pathlib.Path(__file__).resolve().parents[2] / 'lib'))
+
+import configured_logger
+
+logger = configured_logger.new_logger("stderr", stderr=True)
+
+
+def json_rpc(method, params, url):
+    try:
+        j = {
+            'method': method,
+            'params': params,
+            'id': 'dontcare',
+            'jsonrpc': '2.0'
+        }
+        start_time = time.time()
+        r = requests.post(url, json=j, timeout=5)
+        latency_ms = (time.time() - start_time)
+        print(
+            f'prober_request_latency_ms{{method="{method}",url="{url}"}} {latency_ms:.2f}'
+        )
+        result = json.loads(r.content)
+        return result
+    except Exception as e:
+        logger.error(f'Query failed: {e}')
+        sys.exit(1)
+
+
+def get_genesis_height(url):
+    try:
+        genesis_config = json_rpc('EXPERIMENTAL_genesis_config', None, url)
+        genesis_height = genesis_config['result']['genesis_height']
+        logger.info(f'Got genesis_height {genesis_height}')
+        return genesis_height
+    except Exception as e:
+        logger.error(f'get_genesis_height() failed: {e}')
+        sys.exit(1)
+
+
+def get_head(url):
+    try:
+        status = json_rpc('status', None, url)
+        head = status['result']['sync_info']['latest_block_height']
+        logger.info(f'Got latest_block_height {head}')
+        return head
+    except Exception as e:
+        logger.error(f'get_head() failed: {e}')
+        sys.exit(1)
+
+
+def get_block(height, url):
+    try:
+        block = json_rpc('block', {'block_id': height}, url)
+        block = block['result']
+        logger.info(f'Got block at height {height}')
+        return block
+    except Exception as e:
+        logger.error(f'get_block({height}) failed: {e}')
+        return None
+
+
+def get_chunk(chunk, url):
+    try:
+        shard_id = chunk['shard_id']
+        chunk_hash = chunk['chunk_hash']
+        chunk = json_rpc('chunk', {'chunk_id': chunk_hash}, url)
+        chunk = chunk['result']
+        logger.info(f'Got chunk {chunk_hash} for shard {shard_id}')
+        return chunk
+    except Exception as e:
+        logger.error(f'get_chunk({chunk_hash}) failed: {e}')
+        sys.exit(1)
